+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/EncryptedSynchronizedRealmTests.java
@@ -11,6 +11,7 @@ import org.junit.rules.Timeout;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
+import io.realm.BaseIntegrationTest;
 import io.realm.ObjectServerError;
 import io.realm.Realm;
 import io.realm.RealmResults;
@@ -38,16 +39,6 @@ public class EncryptedSynchronizedRealmTests extends BaseIntegrationTest {
 
     @Rule
     public final TestSyncConfigurationFactory configurationFactory = new TestSyncConfigurationFactory();
-
-    @Before
-    public void before() {
-        // This will set the 'm_metadata_manager' in 'sync_manager.cpp' to be 'null'
-        // causing the SyncUser to remain in memory.
-        // They're actually not persisted into disk.
-        // move this call to 'tearDown' to clean in-memory & on-disk users
-        // once https://github.com/realm/realm-object-store/issues/207 is resolved
-        SyncTestUtils.resetSyncMetadata();
-    }
 
     // Make sure the encryption is local, i.e after deleting a synced Realm
     // re-open it again with no (or different) key, should be possible.